Add virtual cursor mixin and make the virtual cursor move like the real cursor
