+++ common/src/main/resources/radium.mixins.json
@@ -6,6 +6,7 @@
   "compatibilityLevel": "JAVA_17",
   "client": [
     "core.MixinFramebuffer",
+    "core.MixinMinecraftDrawVirtualCursor",
     "core.MixinMinecraftFixEarlyCrashNoReports",
     "core.MixinScreenFixClipboard",
     "core.culling.BlockEntityRenderDispatcherMixin",
